+++ internal/interpreter/zz_interpreter_normal_inverse_test.go
@@ -31,6 +31,7 @@ func (f *FakeVideoDevice) PrintChar(r rune)     {}
 func (f *FakeVideoDevice) PrintString(s string) {}
 func (f *FakeVideoDevice) SetCursorX(x int)     {}
 func (f *FakeVideoDevice) SetCursorY(y int)     {}
+func (f *FakeVideoDevice) SwitchMode(slot int)  {}
 func (f *FakeVideoDevice) Plot(x, y int)        {}
 func (f *FakeVideoDevice) ReadLine() (string, error) {
 	return "", nil